refactor(routes): extract upload serving route
