+++ docs/ARCHITECTURE_PHASE_1_ROADMAP.md
@@ -132,6 +132,9 @@ Progress:
   `app/relaytv_app/routes/snapshots.py`.
 - Seeded `app/relaytv_app/routes/ui.py` with the `/` to `/ui` redirect ahead
   of the larger `/ui` extraction.
+- Extracted `GET /media/uploads/{upload_id}/{filename}` into
+  `app/relaytv_app/routes/uploads.py`; ingest/upload playback endpoints remain
+  in the aggregate router until queue/playback extraction.
 
 ### M3: Extract Queue And Playback Routers
 
@@ -263,6 +266,7 @@ Add entries here as PRs land into `codex/architecture-phase-1`.
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by moving thumbnail serving into the asset router. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with snapshots/uploads or simple root redirect. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting snapshot serving and capture aliases. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with uploads or simple root redirect. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by seeding the UI router with the root redirect. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with uploads or capability endpoints. |
+| 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting static uploaded-media serving while leaving ingest/playback upload routes in place. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with capability endpoints or begin planning M3 queue/playback moves. |
 
 ## Open Questions
 
@@ -272,6 +276,6 @@ Add entries here as PRs land into `codex/architecture-phase-1`.
 
 ## Current Recommendation
 
-Continue M2 with low-risk uploads or capability endpoints. Avoid moving
-playback, settings, and Jellyfin routes until the router aggregation pattern is
-proven.
+Continue M2 with capability endpoints or begin planning M3 queue/playback
+moves. Avoid moving settings and Jellyfin routes until the router aggregation
+pattern is proven.
